+++ braintree/transaction.py
@@ -170,7 +170,9 @@ class Transaction(Resource):
         * braintree.Transaction.GatewayRejectionReason.Cvv
         * braintree.Transaction.GatewayRejectionReason.Duplicate
         * braintree.Transaction.GatewayRejectionReason.Fraud
+        * braintree.Transaction.GatewayRejectionReason.RiskThreshold
         * braintree.Transaction.GatewayRejectionReason.ThreeDSecure
+        * braintree.Transaction.GatewayRejectionReason.TokenIssuance
         """
         ApplicationIncomplete = "application_incomplete"
         Avs                   = "avs"
@@ -178,6 +180,7 @@ class Transaction(Resource):
         Cvv                   = "cvv"
         Duplicate             = "duplicate"
         Fraud                 = "fraud"
+        RiskThreshold         = "risk_threshold"
         ThreeDSecure          = "three_d_secure"
         TokenIssuance         = "token_issuance"